refactor: move `ChangeManager` out of `libanki`
com.ichi2.anki.observability.ChangeManager

There was a crash reporting reference, and the implementation differs
from upstream

Issue 18015
