+++ AnkiDroid/src/main/java/com/ichi2/anki/DayRolloverHandler.kt
@@ -35,7 +35,7 @@ import androidx.core.content.ContextCompat.RECEIVER_EXPORTED
 import anki.collection.OpChanges
 import anki.collection.opChanges
 import com.ichi2.anki.CollectionManager.withOpenColOrNull
-import com.ichi2.libanki.ChangeManager
+import com.ichi2.anki.observability.ChangeManager
 import com.ichi2.libanki.EpochSeconds
 import com.ichi2.libanki.sched.Scheduler
 import kotlinx.coroutines.Dispatchers
