+++ AnkiDroid/src/main/java/com/ichi2/anki/observability/OpChangesNotifier.kt
@@ -18,7 +18,7 @@
 package com.ichi2.anki.observability
 
 import com.ichi2.anki.CollectionManager.withCol
-import com.ichi2.libanki.ChangeManager
+import com.ichi2.anki.observability.ChangeManager
 import com.ichi2.libanki.Collection
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.withContext
