+++ AnkiDroid/src/test/java/com/ichi2/testutils/JvmTest.kt
@@ -20,7 +20,7 @@ import android.annotation.SuppressLint
 import androidx.annotation.CallSuper
 import com.ichi2.anki.CollectionManager
 import com.ichi2.anki.common.time.TimeManager
-import com.ichi2.libanki.ChangeManager
+import com.ichi2.anki.observability.ChangeManager
 import com.ichi2.libanki.Collection
 import com.ichi2.libanki.Storage
 import kotlinx.coroutines.Dispatchers
