+++ AnkiDroid/src/test/java/com/ichi2/testutils/TestChangeSubscriber.kt
@@ -17,7 +17,7 @@
 package com.ichi2.testutils
 
 import anki.collection.OpChanges
-import com.ichi2.libanki.ChangeManager
+import com.ichi2.anki.observability.ChangeManager
 import timber.log.Timber
 import kotlin.test.fail
